more emp components
